Removed lots of code that is no longer needed.
